refactor: remove cross fetch name
